backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-ipware==7.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
ipapi==1.0.4
PyJWT==2.8.0
python-ipware==3.0.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
